requirements.txt
pdfplumber
pypdf
python-docx
spacy
